TodoList/requirements.txt
Django==1.8.6
django-cors-headers==2.4.0
djangorestframework==3.4.4
